Optimized Websocket connection and implemented cleanup + Reworked Gameboard and Queue Models + Finished Game Page

+++ api/src/routes/play.ts
@@ -1,15 +1,21 @@
 import express, {Router, Request, Response} from "express";
 import passport from "passport";
 import { Game } from "../models/game.js";
-import { GameQueue } from "../models/queue.js";
+import { GameQueue, IGameQueue } from "../models/queue.js";
 import { createGame, move } from "../game/gamehandler.js";
-import { Model } from "mongoose";
 
 export const PlayRouter: Router = express.Router();
 
 PlayRouter.post('/queue',
   passport.authenticate('jwt', { session: false }),
   async (req: any, res: Response) => {
+
+    if (!req.user || !req.user._id || !req.user.username) {
+      return res.status(400).json({
+        message: "Error adding player to queue",
+        error: "Invalid user data"
+      });
+    }
     try {
       // Check if player was already in the queue -> remove the player from the queue
       if (await GameQueue.findOneAndDelete({ user_id: req.user._id })) {
@@ -20,15 +26,18 @@ PlayRouter.post('/queue',
 
       // Add player to the queue
       await new GameQueue({
-        user_id: req.user._id, username: req.user.username 
+        user_id: req.user._id, 
+        username: req.user.username,
+        ranking: req.user.ranking,
+        title: req.user.title,
       }).save();
 
       // Check if a Game can be created
       // TODO: here could be an algorithmus to find an optimal enemy for the player
       if (await GameQueue.countDocuments() >= 2) {
-        const p1: any = await GameQueue.findOneAndDelete();
-        const p2: any = await GameQueue.findOneAndDelete();
-        await createGame(p1.user_id, p2.user_id, p1.username, p2.username, 20);
+        const p1: IGameQueue | null = await GameQueue.findOneAndDelete();
+        const p2: IGameQueue | null = await GameQueue.findOneAndDelete();
+        await createGame(p1!, p2!, 20);
         // TODO: Maybe redirect user to the board
         return res.status(200).json({
           message: "Created Game",
@@ -62,38 +71,41 @@ PlayRouter.get('/queue', async (req, res) => {
   }
 });
 
-
 PlayRouter.post('/move',
   passport.authenticate('jwt', { session: false }),
   async (req: any, res: Response) => {
 
-    let game;
+    const gameid = req.query.gameid;
+    const discover_id = req.body.discover_id;
+
+    const game = await Game.findById(gameid);
+
+    if (!game) {
+      return res.status(404).json({
+        message: "Error on move",
+        error: `Game with id: ${gameid} not found`
+      });
+    }
+    if (!req.user || (game.player1.id!=req.user._id && game.player2.id!=req.user._id)) {
+      return res.status(403).json({
+        message: "Error on move",
+        error: "Spectators are not allowed to move"
+      });
+    }
+    if (game.game_stage === -1) {
+      return res.status(400).json({
+        message: "Error on move",
+        error: `Game has finished and is now readonly`
+      });
+    }
+    if (game.active_id != req.user._id) {
+      return res.status(403).json({
+        message: "Error on move",
+        error: `You are not allowed to move now`
+      });
+    }
 
     try {
-      const gameid = req.query.gameid;
-      const discover_id = req.body.discover_id;
-
-      game = await Game.findById(gameid);
-
-      if (!game) {
-        return res.status(404).json({
-          message: "Error on move",
-          error: `Game with id: ${gameid} not found`
-        });
-      }
-      if (game.game_stage === -1) {
-        return res.status(400).json({
-          message: "Error on move",
-          error: `Game has finished and is now readonly`
-        });
-      }
-      if (game.active_id !== req.user._id.toString()) {
-        return res.status(403).json({
-          message: "Error on move",
-          error: `You are not allowed to move now`
-        });
-      } 
-
       // This will lock the active_id to prevent a spamming attack
       // The active_id will be re-set in the move() function below
       const active_id_buf = game.active_id;
@@ -103,7 +115,7 @@ PlayRouter.post('/move',
       game.active_id = active_id_buf;
       
       // This will handle the main logic of the programm
-      const enemy_id = game.active_id===game.p1_id ? game.p2_id : game.p1_id;
+      const enemy_id = game.active_id===game.player1.id ? game.player2.id : game.player1.id;
       await move(game, enemy_id, discover_id);
 
       return res.status(200).json({
